python/school-timetabling/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
name = "school_timetabling"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
    'timefold-solver == 999-dev0',
    'fastapi == 0.111.0',
    'pydantic == 2.7.3',
    'uvicorn == 0.30.1',
    'pytest == 8.2.2',
]


[project.scripts]
run-app = "school_timetabling:main"
